+++ tests/PhpWord/Writer/HTML/ElementTest.php
@@ -11,13 +11,15 @@
  * contributors, visit https://github.com/PHPOffice/PHPWord/contributors.
  *
  * @see         https://github.com/PHPOffice/PHPWord
- * @copyright   2010-2017 PHPWord contributors
+ * @copyright   2010-2018 PHPWord contributors
  * @license     http://www.gnu.org/licenses/lgpl.txt LGPL version 3
  */
 
 namespace PhpOffice\PhpWord\Writer\HTML;
 
 use PhpOffice\PhpWord\Element\Text as TextElement;
+use PhpOffice\PhpWord\Element\TrackChange;
+use PhpOffice\PhpWord\PhpWord;
 use PhpOffice\PhpWord\Writer\HTML;
 use PhpOffice\PhpWord\Writer\HTML\Element\Text;
 
@@ -53,5 +55,26 @@ class ElementTest extends \PHPUnit\Framework\TestCase
         $object->setWithoutP(true);
 
         $this->assertEquals(htmlspecialchars('-A-', ENT_COMPAT, 'UTF-8'), $object->write());
+    }
+
+    /**
+     * Test write TrackChange
+     */
+    public function testWriteTrackChanges()
+    {
+        $phpWord = new PhpWord();
+        $section = $phpWord->addSection();
+        $text = $section->addText('my dummy text');
+        $text->setChangeInfo(TrackChange::INSERTED, 'author name');
+        $text2 = $section->addText('my other text');
+        $text2->setTrackChange(new TrackChange(TrackChange::DELETED, 'another author', new \DateTime()));
+
+        $htmlWriter = new HTML($phpWord);
+        $dom = new \DOMDocument();
+        $dom->loadHTML($htmlWriter->getContent());
+        $xpath = new \DOMXpath($dom);
+
+        $this->assertTrue($xpath->query('/html/body/p[1]/ins')->length == 1);
+        $this->assertTrue($xpath->query('/html/body/p[2]/del')->length == 1);
     }
 }