+++ tests/PhpWord/Writer/Word2007/Part/SettingsTest.php
@@ -11,7 +11,7 @@
  * contributors, visit https://github.com/PHPOffice/PHPWord/contributors.
  *
  * @see         https://github.com/PHPOffice/PHPWord
- * @copyright   2010-2017 PHPWord contributors
+ * @copyright   2010-2018 PHPWord contributors
  * @license     http://www.gnu.org/licenses/lgpl.txt LGPL version 3
  */
 
@@ -174,7 +174,7 @@ class SettingsTest extends \PHPUnit\Framework\TestCase
         $this->assertTrue($doc->elementExists($path, $file));
         $element = $doc->getElement($path, $file);
 
-        $this->assertNotEquals('false', $element->getAttribute('w:val'));
+        $this->assertSame('true', $element->getAttribute('w:val'));
     }
 
     /**
@@ -193,7 +193,23 @@ class SettingsTest extends \PHPUnit\Framework\TestCase
         $this->assertTrue($doc->elementExists($path, $file));
 
         $element = $doc->getElement($path, $file);
-        $this->assertNotEquals('false', $element->getAttribute('w:val'));
+        $this->assertSame('true', $element->getAttribute('w:val'));
+    }
+
+    public function testUpdateFields()
+    {
+        $phpWord = new PhpWord();
+        $phpWord->getSettings()->setUpdateFields(true);
+
+        $doc = TestHelperDOCX::getDocument($phpWord);
+
+        $file = 'word/settings.xml';
+
+        $path = '/w:settings/w:updateFields';
+        $this->assertTrue($doc->elementExists($path, $file));
+
+        $element = $doc->getElement($path, $file);
+        $this->assertSame('true', $element->getAttribute('w:val'));
     }
 
     /**
@@ -247,7 +263,7 @@ class SettingsTest extends \PHPUnit\Framework\TestCase
         $this->assertTrue($doc->elementExists($path, $file));
 
         $element = $doc->getElement($path, $file);
-        $this->assertNotEquals('false', $element->getAttribute('w:val'));
+        $this->assertSame('true', $element->getAttribute('w:val'));
     }
 
     /**
@@ -274,6 +290,38 @@ class SettingsTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals('true', $element->getAttribute('w:comments'));
     }
 
+    public function testHideSpellingErrors()
+    {
+        $phpWord = new PhpWord();
+        $phpWord->getSettings()->setHideSpellingErrors(true);
+
+        $doc = TestHelperDOCX::getDocument($phpWord);
+
+        $file = 'word/settings.xml';
+
+        $path = '/w:settings/w:hideSpellingErrors';
+        $this->assertTrue($doc->elementExists($path, $file));
+
+        $element = $doc->getElement($path, $file);
+        $this->assertSame('true', $element->getAttribute('w:val'));
+    }
+
+    public function testHideGrammaticalErrors()
+    {
+        $phpWord = new PhpWord();
+        $phpWord->getSettings()->setHideGrammaticalErrors(true);
+
+        $doc = TestHelperDOCX::getDocument($phpWord);
+
+        $file = 'word/settings.xml';
+
+        $path = '/w:settings/w:hideGrammaticalErrors';
+        $this->assertTrue($doc->elementExists($path, $file));
+
+        $element = $doc->getElement($path, $file);
+        $this->assertSame('true', $element->getAttribute('w:val'));
+    }
+
     /**
      * Test track Revisions
      */
@@ -290,7 +338,7 @@ class SettingsTest extends \PHPUnit\Framework\TestCase
         $this->assertTrue($doc->elementExists($path, $file));
 
         $element = $doc->getElement($path, $file);
-        $this->assertNotEquals('false', $element->getAttribute('w:val'));
+        $this->assertSame('true', $element->getAttribute('w:val'));
     }
 
     /**
@@ -309,7 +357,7 @@ class SettingsTest extends \PHPUnit\Framework\TestCase
         $this->assertTrue($doc->elementExists($path, $file));
 
         $element = $doc->getElement($path, $file);
-        $this->assertNotEquals('false', $element->getAttribute('w:val'));
+        $this->assertSame('true', $element->getAttribute('w:val'));
     }
 
     /**
@@ -328,6 +376,70 @@ class SettingsTest extends \PHPUnit\Framework\TestCase
         $this->assertTrue($doc->elementExists($path, $file));
 
         $element = $doc->getElement($path, $file);
-        $this->assertNotEquals('false', $element->getAttribute('w:val'));
+        $this->assertSame('true', $element->getAttribute('w:val'));
+    }
+
+    public function testAutoHyphenation()
+    {
+        $phpWord = new PhpWord();
+        $phpWord->getSettings()->setAutoHyphenation(true);
+
+        $doc = TestHelperDOCX::getDocument($phpWord);
+
+        $file = 'word/settings.xml';
+
+        $path = '/w:settings/w:autoHyphenation';
+        $this->assertTrue($doc->elementExists($path, $file));
+
+        $element = $doc->getElement($path, $file);
+        $this->assertSame('true', $element->getAttribute('w:val'));
+    }
+
+    public function testConsecutiveHyphenLimit()
+    {
+        $phpWord = new PhpWord();
+        $phpWord->getSettings()->setConsecutiveHyphenLimit(2);
+
+        $doc = TestHelperDOCX::getDocument($phpWord);
+
+        $file = 'word/settings.xml';
+
+        $path = '/w:settings/w:consecutiveHyphenLimit';
+        $this->assertTrue($doc->elementExists($path, $file));
+
+        $element = $doc->getElement($path, $file);
+        $this->assertSame('2', $element->getAttribute('w:val'));
+    }
+
+    public function testHyphenationZone()
+    {
+        $phpWord = new PhpWord();
+        $phpWord->getSettings()->setHyphenationZone(100);
+
+        $doc = TestHelperDOCX::getDocument($phpWord);
+
+        $file = 'word/settings.xml';
+
+        $path = '/w:settings/w:hyphenationZone';
+        $this->assertTrue($doc->elementExists($path, $file));
+
+        $element = $doc->getElement($path, $file);
+        $this->assertSame('100', $element->getAttribute('w:val'));
+    }
+
+    public function testDoNotHyphenateCaps()
+    {
+        $phpWord = new PhpWord();
+        $phpWord->getSettings()->setDoNotHyphenateCaps(true);
+
+        $doc = TestHelperDOCX::getDocument($phpWord);
+
+        $file = 'word/settings.xml';
+
+        $path = '/w:settings/w:doNotHyphenateCaps';
+        $this->assertTrue($doc->elementExists($path, $file));
+
+        $element = $doc->getElement($path, $file);
+        $this->assertSame('true', $element->getAttribute('w:val'));
     }
 }