+++ CHANGELOG.md
@@ -13,7 +13,11 @@ v0.15.0 (?? ??? 2018)
 - Add support for fixed Table Layout @aoloe @ekopach @troosan #841 #1276
 - Add support for Cell Spacing @dox07 @troosan #1040
 - Add parsing of formatting inside lists @atomicalnet @troosan #594
+- Added support for Vertically Raised or Lowered Text (w:position) @anrikun @troosan #640
 - Add support for MACROBUTTON field @phryneas @troosan #1021
+- Add support for Hyphenation @Trainmaster #1282 (Document: `autoHyphenation`, `consecutiveHyphenLimit`, `hyphenationZone`, `doNotHyphenateCaps`, Paragraph: `suppressAutoHyphens`)
+- Added support for Floating Table Positioning (tblpPr) @anrikun #639
+- Added support for Image text wrapping distance @troosan #1310
 
 ### Fixed
 - Fix reading of docx default style - @troosan #1238
@@ -23,6 +27,7 @@ v0.15.0 (?? ??? 2018)
 - Fix parsing of `<w:br/>` tag. @troosan #1274
 - Bookmark are not writton as internal link in html writer @troosan #1263
 - It should be possible to add a Footnote in a ListItemRun @troosan #1287 #1287
+- Fix parsing of Heading and Title formating @troosan @gthomas2 #465
 
 ### Changed
 - Remove zend-stdlib dependency @Trainmaster #1284